update test_junit5 DucNV 27/12
